axes are fully responsible for calculating tickValues and tickFormat
