+++ test/client/spec/components/victory-chart/helper-methods.spec.js
@@ -157,107 +157,4 @@ describe("victory-chart/helpers-methods", () => {
       expect(stringResult).to.eql({ a: 1, b: 2, c: 3, d: 4 });
     });
   });
-
-  describe("getTickFormat", () => {
-    const stringMap = { x: { "a": 1, "b": 2, "c": 3 } };
-    const nullStringMap = { x: null };
-
-    it("returns the identity function when tickValues are numerical", () => {
-      const props = { tickValues: [1, 2, 3] };
-      const victoryAxis = getVictoryAxis(props);
-      const formatResult = Helpers.getTickFormat(victoryAxis, "x", { stringMap: nullStringMap });
-      const val = 2;
-      expect(formatResult).to.be.a("function");
-      expect(formatResult(val)).to.equal(val);
-    });
-
-    it("returns a function from a string map", () => {
-      const victoryAxis = getVictoryAxis({});
-      const formatResult = Helpers.getTickFormat(victoryAxis, "x", { stringMap });
-      expect(formatResult).to.be.a("function");
-      expect(formatResult(1)).to.equal("a");
-    });
-
-    it("returns undefined if no tickValues or stringMaps exist", () => {
-      const victoryAxis = getVictoryAxis({});
-      const formatResult = Helpers.getTickFormat(
-        victoryAxis, "x", { stringMap: nullStringMap }
-      );
-      expect(formatResult).to.equal(undefined);
-    });
-  });
-
-  describe("getTicks", () => {
-    let sandbox;
-    beforeEach(() => {
-      sandbox = sinon.sandbox.create();
-      const fakeTicks = (bool) => bool;
-      sandbox.stub(Helpers, "getTicksFromAxis", fakeTicks);
-      sandbox.stub(Helpers, "getTicksFromData", fakeTicks);
-    });
-
-    afterEach(() => {
-      sandbox.restore();
-    });
-
-    it("returns the result of getTicksFromAxis if truthy", () => {
-      expect(Helpers.getTicks(true)).to.eql(true);
-      expect(Helpers.getTicksFromAxis).calledWith(true).and.returned(true);
-      expect(Helpers.getTicksFromData).notCalled;
-    });
-
-    it("returns the result of getTicksFromData if getTicksFromAxis didnt return ticks", () => {
-      expect(Helpers.getTicks(false)).to.eql(false);
-      expect(Helpers.getTicksFromAxis).calledWith(false).and.returned(false);
-      expect(Helpers.getTicksFromData).calledWith(false).and.returned(false);
-    });
-  });
-
-  describe("getTicksFromAxis", () => {
-    const stringMap = { x: { "a": 1, "b": 2, "c": 3 } };
-
-    it("returns numeric tickValues from a child component props", () => {
-      const props = { tickValues: [1, 2, 3] };
-      const victoryAxis = getVictoryAxis(props);
-      const tickResult = Helpers.getTicksFromAxis({ stringMap }, "x", victoryAxis);
-      expect(tickResult).to.eql(props.tickValues);
-    });
-
-    it("converts string tickValues to numbers", () => {
-      const props = { tickValues: ["a", "b", "c"] };
-      const victoryAxis = getVictoryAxis(props);
-      const tickResult = Helpers.getTicksFromAxis({ stringMap }, "x", victoryAxis);
-      expect(tickResult).to.eql([1, 2, 3]);
-    });
-
-    it("converts returns undefined if tickValues is not given", () => {
-      const victoryAxis = getVictoryAxis({});
-      const tickResult = Helpers.getTicksFromAxis({ stringMap }, "x", victoryAxis);
-      expect(tickResult).to.be.undefined;
-    });
-  });
-
-  describe("getTicksFromData", () => {
-
-    it("returns a category array if there is one", () => {
-      const categories = { x: [1, 2, 3] };
-      const stringMap = { x: { "a": 1, "b": 2, "c": 3 } };
-      const tickResult = Helpers.getTicksFromData({ categories, stringMap }, "x");
-      expect(tickResult).to.eql(categories.x);
-    });
-
-    it("returns the values from a string map if categories are not given", () => {
-      const stringMap = { x: { "a": 1, "b": 2, "c": 3 } };
-      const categories = {};
-      const tickResult = Helpers.getTicksFromData({ categories, stringMap }, "x");
-      expect(tickResult).to.eql([1, 2, 3]);
-    });
-
-    it("returns undefined if neither a string map or categories are given", () => {
-      const stringMap = {};
-      const categories = {};
-      const tickResult = Helpers.getTicksFromData({ categories, stringMap }, "x");
-      expect(tickResult).to.be.undefined;
-    });
-  });
 });
